+++ pkg/vault/dummy.go
@@ -1,7 +1,5 @@
-// +build testutils
-
 /*
-Copyright 2017 Tuenti Technologies S.L. All rights reserved.
+Copyright 2018 Tuenti Technologies S.L. All rights reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -22,57 +20,8 @@ import (
 	"path"
 	"testing"
 
-	log "github.com/mgutz/logxi/v1"
-
 	"github.com/hashicorp/vault/api"
-	"github.com/hashicorp/vault/builtin/credential/approle"
-	"github.com/hashicorp/vault/helper/logformat"
-	"github.com/hashicorp/vault/logical"
-	"github.com/hashicorp/vault/physical"
-	v "github.com/hashicorp/vault/vault"
 )
-
-// Based on TestCores on github.com/hashicorp/vault
-func NewTestCoreAppRole(t *testing.T) (*v.Core, [][]byte, string) {
-	logLevel := log.LevelError
-	if testing.Verbose() {
-		logLevel = log.LevelTrace
-	}
-	logger := logformat.NewVaultLogger(logLevel)
-	physicalBackend := physical.NewInmem(logger)
-
-	credentialBackends := make(map[string]logical.Factory)
-	credentialBackends["approle"] = approle.Factory
-
-	conf := &v.CoreConfig{
-		Physical:           physicalBackend,
-		CredentialBackends: credentialBackends,
-		DisableMlock:       true,
-		Logger:             logger,
-	}
-	core, err := v.NewCore(conf)
-
-	keys, token := v.TestCoreInit(t, core)
-	for _, key := range keys {
-		if _, err := v.TestCoreUnseal(core, v.TestKeyCopy(key)); err != nil {
-			t.Fatalf("unseal err: %s", err)
-		}
-	}
-
-	sealed, err := core.Sealed()
-	if err != nil {
-		t.Fatalf("err checking seal status: %s", err)
-	}
-	if sealed {
-		t.Fatal("should not be sealed")
-	}
-
-	if err != nil {
-		t.Fatalf("couldn't start core: %v", err)
-	}
-
-	return core, keys, token
-}
 
 func setupAppRole(t *testing.T, name, token, address string, secret bool) string {
 	v := vaultApi{
